CDK-128: Adding NoSuchDatasetException.
Callers should be able to handle a missing dataset. NoSuchDatasetException will
be thrown when an operation is requested on a data set or descriptor that is
missing---except for delete, which returns false. This also fixes a problem in
the AbstractMetadataProvider's save implementation, where it had to assume that
any MetadataProviderException meant that the descriptor was missing. Now, it
catches just the NoSuchDatasetException and correctly calls either create or
update.